pyproject.toml
[project]
name = "summarizer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.8",
]
